+++ plop/general/component-stories-mdx.txt
@@ -1,6 +1,6 @@
 import {{properCase componentName}} from "../{{properCase componentName}}";
 import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import styles from "./{{properCase componentName}}.stories.module.scss";
 
 export const metaSettings = createStoryMetaSettings({
