+++ src/components/AlertBanner/__stories__/alertBanner.stories.mdx
@@ -4,7 +4,7 @@ import AlertBannerText from "../AlertBannerText/AlertBannerText";
 import AlertBannerLink from "../AlertBannerLink/AlertBannerLink";
 import AlertBannerButton from "../AlertBannerButton/AlertBannerButton";
 import { ComponentRules, Link, RelatedComponents, UsageGuidelines } from "../../../storybook/components";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import {
   ATTENTION_BOX,
   TOAST,
