+++ src/components/Avatar/__stories__/avatar.stories.mdx
@@ -1,5 +1,5 @@
 import Avatar from "../Avatar";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import {
   AVATAR_GROUP,
