+++ src/components/AvatarGroup/__stories__/AvatarGroup.stories.mdx
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import { StoryDescription } from "../../../storybook/components/story-description/story-description";
 import { Link, RelatedComponents, UsageGuidelines } from "../../../storybook";
 import { AVATAR, COUNTER, TOOLTIP } from "../../../storybook/components/related-components/component-description-map";
