+++ src/components/Box/__stories__/Box.stories.mdx
@@ -1,6 +1,6 @@
 import { Box } from "../../../components";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import { RelatedComponents, Tip, UsageGuidelines } from "../../../storybook";
 import {
   DIALOG_CONTENT_CONTAINER,
