+++ src/components/BreadcrumbsBar/BreadcrumbItem/__stories__/breadcrumbItem.stories.mdx
@@ -1,8 +1,5 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import {
-  createComponentTemplate,
-  createStoryMetaSettings
-} from "../../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../../storybook";
 import BreadcrumbsBar from "../../BreadcrumbsBar";
 import BreadcrumbItem from "../BreadcrumbItem";
 import { Workspace } from "../../../Icon/Icons";
