+++ src/components/BreadcrumbsBar/__stories__/breadcrumbsBar.stories.mdx
@@ -7,7 +7,7 @@ import { Board, Folder, Group, Workspace } from "../../Icon/Icons";
 import person3 from "./assets/person3.png";
 import { Link } from "../../../storybook/components";
 import { BUTTON_GROUP, TABS, WIZARD } from "../../../storybook/components/related-components/component-description-map";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import "./breadcrumbsBar.stories.scss";
 
 export const metaSettings = createStoryMetaSettings({
