+++ src/components/ColorPicker/__stories__/ColorPicker.stories.mdx
@@ -1,16 +1,16 @@
 import ColorPicker from "../ColorPicker";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../storybook";
 import {
-  colorPickerTemplate,
-  colorPickerWithIndicatorTemplate,
-  colorPickerTextIndicatorTemplate,
-  colorPickerSelectedTemplate,
+  colorPickerCircleShapeTemplate,
   colorPickerNoColorTemplate,
   colorPickerSelectedIconTemplate,
-  colorPickerCircleShapeTemplate
+  colorPickerSelectedTemplate,
+  colorPickerTemplate,
+  colorPickerTextIndicatorTemplate,
+  colorPickerWithIndicatorTemplate
 } from "./ColorPicker.stories";
-import { noColorInteractionSuite, multiSelectionInteractionSuite } from "../__tests__/ColorPicker.interactions.js"
+import { multiSelectionInteractionSuite, noColorInteractionSuite } from "../__tests__/ColorPicker.interactions.js";
 
 export const metaSettings = createStoryMetaSettings({
   component: ColorPicker,
