+++ src/components/Counter/__stories__/counter.stories.mdx
@@ -1,7 +1,7 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { useCallback, useEffect, useState } from "react";
 import Counter from "../Counter";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import { CHIP, LABEL, TOOLTIP } from "../../../storybook/components/related-components/component-description-map";
 import Avatar from "../../Avatar/Avatar";
 import { AddUpdate, Notifications, Update } from "../../Icon/Icons";
