+++ src/components/DialogContentContainer/__stories__/DialogContentContainer.stories.mdx
@@ -1,6 +1,6 @@
 import DialogContentContainer from "../DialogContentContainer";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import DialogContentContainerExample from "./DialogContentContainerExample";
 
 export const metaSettings = createStoryMetaSettings({
