+++ src/components/Divider/__stories__/Divider.stories.mdx
@@ -1,6 +1,6 @@
 import Divider from "../Divider";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../storybook";
 import { CHIP, ICONS, LABEL } from "../../../storybook/components/related-components/component-description-map";
 import classes from "./Divider.stories.module.scss";
