+++ src/components/EditableHeading/__stories__/EditableHeading.stories.mdx
@@ -1,6 +1,6 @@
 import EditableHeading from "../EditableHeading";
-import {ArgsTable, Story, Canvas, Meta} from "@storybook/addon-docs";
-import {createComponentTemplate, createStoryMetaSettings} from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 
 export const metaSettings = createStoryMetaSettings({
     component: EditableHeading,
