+++ src/components/GridKeyboardNavigationContext/__stories__/useGridKeyboardNavigationContext.stories.mdx
@@ -1,10 +1,10 @@
 import { useRef } from "react";
-import { useGridKeyboardNavigationContext, GridKeyboardNavigationContext } from "../GridKeyboardNavigationContext";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { GridKeyboardNavigationContext, useGridKeyboardNavigationContext } from "../GridKeyboardNavigationContext";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import { DummyNavigableGrid, LayoutWithInnerKeyboardNavigation } from "./useGridKeyboardNavigationContext.stories";
 import { Flex } from "../..";
-import { useGridContextMultipleDepthsPlaySuite } from "../__tests__/useGridKeyboardNavigationContext.interactions"
+import { useGridContextMultipleDepthsPlaySuite } from "../__tests__/useGridKeyboardNavigationContext.interactions";
 
 export const metaSettings = createStoryMetaSettings({
   component: useGridKeyboardNavigationContext
