+++ src/components/Heading/__stories__/Heading.stories.mdx
@@ -1,6 +1,6 @@
 import Heading from "../Heading";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import { Link, RelatedComponents } from "../../../storybook/components";
 import Divider from "../../Divider/Divider";
 import Search from "../../Search/Search";
