+++ src/components/HiddenText/__stories__/HiddenText.stories.mdx
@@ -1,6 +1,6 @@
 import HiddenText from "../HiddenText";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate } from "../../../storybook/functions/createComponentTemplate";
 import {
   CLICKABLE,
   HEADING,
