+++ src/components/Icon/__stories__/Icon.stories.mdx
@@ -1,7 +1,7 @@
 import Icon from "../Icon";
 import Bolt from "../Icons/components/Bolt";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import { IconsList } from "./icon.stories.helperComponents";
 import "./Icon.stories.scss";
