+++ src/components/Label/__stories__/label.stories.mdx
@@ -1,8 +1,8 @@
 import Label from "../Label";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
-import { TOOLTIP, COUNTER, CHIP } from "../../../storybook/components/related-components/component-description-map";
-import { MultipleStoryElementsWrapper, Link } from "../../../storybook/components";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
+import { CHIP, COUNTER, TOOLTIP } from "../../../storybook/components/related-components/component-description-map";
+import { Link, MultipleStoryElementsWrapper } from "../../../storybook/components";
 import "./label.stories.scss";
 
 export const metaSettings = createStoryMetaSettings({
