+++ src/components/ListItemIcon/__stories__/ListItemIcon.stories.mdx
@@ -1,6 +1,6 @@
 import ListItemIcon from "../ListItemIcon";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 
 export const metaSettings = createStoryMetaSettings({
   component: ListItemIcon,
