+++ src/components/Menu/MenuDivider/__stories__/MenuDivider.stories.mdx
@@ -1,6 +1,6 @@
 import MenuDivider from "../MenuDivider";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate } from "../../../../storybook/functions/createComponentTemplate";
 import Menu from "../../Menu/Menu";
 import MenuItem from "../../MenuItem/MenuItem";
