+++ src/components/Menu/MenuGridItem/__stories__/MenuGridItem.stories.mdx
@@ -1,11 +1,8 @@
 import MenuGridItem from "../MenuGridItem";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import {
-  createComponentTemplate,
-  createStoryMetaSettings
-} from "../../../../storybook/functions/create-component-story";
-import { MENU_BUTTON, MENU } from "../../../../storybook/components/related-components/component-description-map";
-import { menuGridItemTemplate, menuGridItemWithDisabled, menuGridItemInSubMenus } from "./MenuGridItem.stories";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../../storybook";
+import { MENU, MENU_BUTTON } from "../../../../storybook/components/related-components/component-description-map";
+import { menuGridItemInSubMenus, menuGridItemTemplate, menuGridItemWithDisabled } from "./MenuGridItem.stories";
 import "./MenuGridItem.stories.scss";
 
 export const metaSettings = createStoryMetaSettings({
