+++ src/components/Menu/MenuItem/__stories__/MenuItem.stories.mdx
@@ -1,11 +1,11 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../../storybook/functions/createStoryMetaSettings";
 import {
-  menuItemTemplate,
-  menuItemStatesTemplate,
   menuItemIconsTemplate,
   menuItemIconsWithColorsTemplate,
   menuItemOverflowTemplate,
+  menuItemStatesTemplate,
+  menuItemTemplate,
   menuItemTooltipTemplate
 } from "./MenuItem.stories";
 import MenuItem from "../MenuItem";
