+++ src/components/Menu/MenuItemButton/__stories__/MenuItemButton.stories.mdx
@@ -1,13 +1,10 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import {
-  createComponentTemplate,
-  createStoryMetaSettings
-} from "../../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../../storybook";
 import MenuItemButton from "../MenuItemButton";
 import {
-  menuItemButtonKindsTemplate,
+  menuItemButtonDisabledTemplate,
   menuItemButtonIconTemplate,
-  menuItemButtonDisabledTemplate
+  menuItemButtonKindsTemplate
 } from "./menuItemButton.stories";
 
 export const metaSettings = createStoryMetaSettings({
