+++ src/components/Menu/MenuTitle/__stories__/MenuTitle.stories.mdx
@@ -1,7 +1,5 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import {
-  createStoryMetaSettings
-} from "../../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createStoryMetaSettings } from "../../../../storybook/functions/createStoryMetaSettings";
 import MenuTitle from "../MenuTitle";
 import { menuTitleCaptionPositionsTemplate, menuTitleTemplate } from "./MenuTitle.stories";
