+++ src/components/MultiStepIndicator/__stories__/multiStepIndicator.stories.mdx
@@ -1,12 +1,12 @@
 import MultiStepIndicator from "../MultiStepIndicator";
-import { useMemo, useState, useEffect, useCallback } from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { STEP_STATUSES } from "../MultiStepConstants";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate } from "../../../storybook/functions/createComponentTemplate";
 import { BREADCRUBMS, STEPPER, TABS } from "../../../storybook/components/related-components/component-description-map";
 import { firstSteps, secondSteps, thirdSteps } from "./helper.js";
 import { Link } from "../../../storybook/components";
-import { Featured } from "../../Icon/Icons"
+import { Featured } from "../../Icon/Icons";
 import "./multiStepIndicator.stories.scss";
