+++ src/components/ProgressBars/LinearProgressBar/__stories__/linearProgressBar.stories.mdx
@@ -1,7 +1,7 @@
 import LinearProgressBar from "../LinearProgressBar";
 import { useMemo } from "react";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../../storybook/functions/createStoryMetaSettings";
 import colors from "../../../../constants/colors.json";
 import DialogContentContainer from "../../../DialogContentContainer/DialogContentContainer";
 import { Info } from "../../../Icon/Icons";
