+++ src/components/RadioButton/__stories__/radioButton.stories.mdx
@@ -1,7 +1,7 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import RadioButton from "../RadioButton";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
-import { TOGGLE, CHECKBOX, DROPDOWN } from "../../../storybook/components/related-components/component-description-map";
+import { createComponentTemplate } from "../../../storybook/functions/createComponentTemplate";
+import { CHECKBOX, DROPDOWN, TOGGLE } from "../../../storybook/components/related-components/component-description-map";
 import DialogContentContainer from "../../DialogContentContainer/DialogContentContainer";
 import { Link } from "../../../storybook/components";
 import { clickRadioButton, controlRadioButton } from "./interactions";
