+++ src/components/ResponsiveList/__stories__/ResponsiveList.stories.mdx
@@ -1,6 +1,5 @@
 import ResponsiveList from "../ResponsiveList";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { Sandbox } from "./responsiveListStoriesComponents";
 import "./ResponsiveList.stories.scss";
