+++ src/components/Search/__stories__/Search.stories.mdx
@@ -1,6 +1,6 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import Search from "../Search";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import DialogContentContainer from "../../DialogContentContainer/DialogContentContainer";
 import { Link, RelatedComponents, Tip, UsageGuidelines } from "../../../storybook/components";
 import Combobox from "../../Combobox/Combobox.jsx";
