+++ src/components/Skeleton/__stories__/skeleton.stories.mdx
@@ -1,16 +1,15 @@
 import Skeleton from "../Skeleton";
-import { useMemo, useState, useEffect, useCallback } from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { useCallback, useState } from "react";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import {
-  PROGRESS_BAR,
   COUNTER,
+  PROGRESS_BAR,
   SPINNER
 } from "../../../storybook/components/related-components/component-description-map";
 import Avatar from "../../Avatar/Avatar";
 import person from "./assets/person.png";
 import Button from "../../Button/Button";
-import { ThumbsUp, Replay } from "../../Icon/Icons";
+import { Replay, ThumbsUp } from "../../Icon/Icons";
 import { Link } from "../../../storybook/components";
 import "./skeleton.stories.scss";
