+++ src/components/SplitButton/__stories__/splitButton.stories.mdx
@@ -1,12 +1,12 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { useMemo } from "react";
 import SplitButton from "../SplitButton";
 import Button from "../../Button/Button";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
-import { Search, Filter, Download, Group, Upload, Print, Sort, Add } from "../../Icon/Icons";
-import { BUTTON_GROUP, BUTTON, MENU } from "../../../storybook/components/related-components/component-description-map";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
+import { Add, Download, Filter, Group, Print, Search, Sort, Upload } from "../../Icon/Icons";
+import { BUTTON, BUTTON_GROUP, MENU } from "../../../storybook/components/related-components/component-description-map";
 import SplitButtonExampleDialog from "./SplitButtonExampleDialog";
-import { MultipleStoryElementsWrapper, Link } from "../../../storybook/components";
+import { Link } from "../../../storybook/components";
 import "./splitButton.stories.scss";
 
 export const metaSettings = createStoryMetaSettings({
