+++ src/components/Steps/__stories__/steps.stories.mdx
@@ -1,8 +1,8 @@
 import Steps from "../Steps";
-import {useCallback, useMemo, useState} from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
-import Tipseen from "../../Tipseen/Tipseen.jsx"
+import { useMemo } from "react";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
+import Tipseen from "../../Tipseen/Tipseen.jsx";
 import TipseenWizard from "../../Tipseen/TipseenWizard";
 import { modifiers } from "./helpers.js";
 import { BREADCRUBMS, TABS, WIZARD } from "../../../storybook/components/related-components/component-description-map";
