+++ src/components/Tabs/Tab/__stories__/tab.stories.mdx
@@ -1,6 +1,6 @@
 import Tab from "../Tab";
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../../storybook/functions/create-component-story";
+import { createComponentTemplate } from "../../../../storybook/functions/createComponentTemplate";
 import { Email } from "../../../Icon/Icons";
 
 <Meta
