+++ src/components/Tabs/TabPanel/__stories__/tab-panel.stories.mdx
@@ -1,5 +1,5 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../../storybook/functions/create-component-story";
+import { createComponentTemplate } from "../../../../storybook/functions/createComponentTemplate";
 import TabPanel from "../TabPanel";
 
 <Meta
