+++ src/components/Tabs/TabPanels/__stories__/tab-panels.stories.mdx
@@ -1,5 +1,5 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { createStoryMetaSettings } from "../../../../storybook/functions/create-component-story";
+import { createStoryMetaSettings } from "../../../../storybook/functions/createStoryMetaSettings";
 import TabPanel from "../../TabPanel/TabPanel";
 import TabPanels from "../TabPanels";
