+++ src/components/TextWithHighlight/__stories__/TextWithHighlight.stories.mdx
@@ -1,6 +1,6 @@
 import TextWithHighlight from "../TextWithHighlight";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+import { createComponentTemplate } from "../../../storybook/functions/createComponentTemplate";
 import {
   EDITABLE_HEADING,
   HEADING,
