+++ src/components/Tipseen/__stories__/tipseen.stories.mdx
@@ -1,15 +1,19 @@
 import Tipseen from "../Tipseen";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import TipseenContent from "../TipseenContent";
 import TipseenWizard from "../TipseenWizard";
 import TipseenImage from "../TipseenImage";
-import picture from "./assets/picture.png"
+import picture from "./assets/picture.png";
 import Icon from "../../Icon/Icon";
-import { BoardPrivate, Board } from "../../Icon/Icons"
+import { Board, BoardPrivate } from "../../Icon/Icons";
 import { modifiers } from "./helper";
 import { Link } from "../../../storybook/components";
-import { TOOLTIP, ATTENTION_BOX, CHIP } from "../../../storybook/components/related-components/component-description-map";
-import { createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import {
+  ATTENTION_BOX,
+  CHIP,
+  TOOLTIP
+} from "../../../storybook/components/related-components/component-description-map";
+import { createStoryMetaSettings } from "../../../storybook/functions/createStoryMetaSettings";
 import { tipseenTemplate } from "./tipseen.stories";
 import "./tipseen.stories.scss";
