+++ src/components/Toast/__stories__/toast.stories.mdx
@@ -1,13 +1,13 @@
-import { useState, useCallback, useMemo } from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { useCallback, useMemo, useState } from "react";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import Toast from "../Toast";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import Button from "../../Button/Button";
 import { Link } from "../../../storybook/components";
 import {
-  TOOLTIP,
+  ALERT_BANNER,
   ATTENTION_BOX,
-  ALERT_BANNER
+  TOOLTIP
 } from "../../../storybook/components/related-components/component-description-map";
 import { Delete } from "../../Icon/Icons";
 import "./toast.stories.scss";
