+++ src/components/Toggle/__stories__/toggle.stories.mdx
@@ -1,6 +1,6 @@
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import Toggle from "../Toggle";
-import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
+import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook";
 import {
   BUTTON_GROUP,
   CHECKBOX,
