+++ src/components/VirtualizedGrid/__stories__/VirtualizedGrid.stories.mdx
@@ -1,8 +1,7 @@
 import { useCallback, useMemo, useState } from "react";
 import VirtualizedGrid from "../VirtualizedGrid";
 import Button from "../../Button/Button";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createComponentTemplate } from "../../../storybook/functions/create-component-story";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { generateItems, itemRenderer } from "./VirtualizedGrid.stories";
 
 <Meta title="Navigation/VirtualizedGrid" component={VirtualizedGrid} />
