+++ src/hooks/useAfterFirstRender/__stories__/useAfterFirstRender.stories.mdx
@@ -1,5 +1,6 @@
-import { Story, Canvas, Meta } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useAfterFirstRender from "..";
+import { FunctionArgument, FunctionArguments, UsageGuidelines } from "../../../storybook";
 
 <Meta title="Hooks/useAfterFirstRender" />
