+++ src/hooks/useClickOutside/__stories__/useClickOutside.stories.mdx
@@ -1,6 +1,7 @@
 import { useCallback, useRef, useState } from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useClickOutside from "..";
+import { FunctionArgument, FunctionArguments, UsageGuidelines } from "../../../storybook";
 import "../../__stories__/general-hooks-stories.scss";
 
 <Meta title="Hooks/useClickOutside" />
