+++ src/hooks/useDebounceEvent/__stories__/useDebounceEvent.stories.mdx
@@ -1,7 +1,8 @@
 import { useState } from "react";
-import { Story, Canvas, Meta, ArgsTable } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useDebounceEvent from "..";
 import "../../__stories__/general-hooks-stories.scss";
+import { FunctionArgument, FunctionArguments } from "../../../storybook";
 
 <Meta title="Hooks/useDebounceEvent" />
