+++ src/hooks/useKeyEvent/__stories__/useKeyEvent.stories.mdx
@@ -1,6 +1,7 @@
 import { useState } from "react";
-import { Story, Canvas, Meta, ArgsTable } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useKeyEvent from "..";
+import { FunctionArgument, FunctionArguments, UsageGuidelines } from "../../../storybook";
 import "../../__stories__/general-hooks-stories.scss";
 
 <Meta title="Hooks/useKeyEvent" />
