+++ src/hooks/useTimeout/__stories__/useTimeout.stories.mdx
@@ -1,6 +1,7 @@
 import { useCallback } from "react";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useTimeout from "..";
+import { FunctionArgument, FunctionArguments } from "../../../storybook";
 
 <Meta title="Hooks/useTimeout" />
