+++ src/hooks/useVibeMediaQuery/__stories__/useVibeMediaQuery.stories.mdx
@@ -1,7 +1,8 @@
-import { Story, Canvas, Meta } from "@storybook/addon-docs";
+import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import useVibeMediaQuery from "..";
 import "../../__stories__/general-hooks-stories.scss";
 import styles from "./useVibeMediaQuery.modules.scss";
+import { FunctionArgument, FunctionArguments } from "../../../storybook";
 
 <Meta title="Hooks/useVibeMediaQuery" />
 
@@ -36,7 +37,7 @@ This hook will return the value of the current vibe break point
 
 ## Arguments
 
-This hook doesn't receives any parameters
+This hook doesn't receive any parameters
 
 ## Returns
