+++ src/storybook/index.js
@@ -1,4 +1,5 @@
 export * from "./components";
 export * from "./decorators/vertical-stories";
 export * from "./decorators/memory-stats";
-export * from "./functions/create-component-story";
+export * from "./functions/createStoryMetaSettings";
+export * from "./functions/createComponentTemplate";
